feat(matcher): E1.3c — move patient_info_attributes + mappers into exact_matching
Third slice of E1.3, and the milestone: the per-trial matcher is now fully
self-contained from trials.* (no top-level trials import remains — only the
TYPE_CHECKING Trial/PatientInfo hints, the DjangoMatcherData default port, and
one deferred soft-dep, see below).

Moves 5 files + shims:
- patient_info/patient_info_attributes.py (960L)
- therapies_mapper.py (1508L, pure)
- trial_details/configs.py (911L)
- user_to_trial_attrs_mapper.py (333L)
- attribute_names.py (61L, pure leaf)

Internal edges repoint to package paths (self-containment). The pre-existing
circular import between patient_info_attributes and user_to_trial_attrs_mapper
is preserved 1:1 — both move together, same import order — and the full suite
confirms it still resolves. inflection declared as a package dep.

Option A (deferred models-ownership, per user): trial_details/configs.py keeps
two function-local, lazy `from trials.models import <Units enums>` as-is — a
soft runtime dep on the host providing those TextChoices names (same shape as
the TYPE_CHECKING Trial/PatientInfo coupling). Settling where the Units enums
live is D1/E1.3d work, not this mechanical slice.

Old paths are import-* shims. External importers unchanged: PIA ×9,
therapies_mapper ×7, user_to_trial_attrs_mapper ×6, attribute_names ×6,
trial_details.configs ×3.

Full EXACT suite green (1073 passed, 5 skipped).

Part of EPIC cancerbot#4601 (E1 #272, D1 resolved). Proposal — cross-team.
